Projects's urls easier to read
Previous urls were: /project/project_short_name
New urls are: /2014/project_short_name

Most of the time people shorten urls in order to get one step up in the hierarchy.
With this new pattern, it is possible to do so.

Moreover, previous urls can still be used. This feature has been kept in an attempt to reduce the effect of this change on the current ranking.